handle fallout of new type indirections

+++ crates/re_types/src/components/draw_order_ext.rs
@@ -1,25 +1,27 @@
+use re_types_core::datatypes::Float32;
+
 use super::DrawOrder;
 
 // TODO(cmc): come up with some DSL in our flatbuffers definitions so that we can declare these
 // constants directly in there.
 impl DrawOrder {
     /// Draw order used for depth image when they're not shown as a 3D point cloud and no draw order was specified.
-    pub const DEFAULT_DEPTH_IMAGE: Self = Self(-20.0);
+    pub const DEFAULT_DEPTH_IMAGE: Self = Self(Float32(-20.0));
 
     /// Draw order used for images if no draw order was specified.
-    pub const DEFAULT_IMAGE: Self = Self(-10.0);
+    pub const DEFAULT_IMAGE: Self = Self(Float32(-10.0));
 
     /// Draw order used for segmentation images if no draw order was specified.
-    pub const DEFAULT_SEGMENTATION_IMAGE: Self = Self(0.0);
+    pub const DEFAULT_SEGMENTATION_IMAGE: Self = Self(Float32(0.0));
 
     /// Draw order used for 2D boxes if no draw order was specified.
-    pub const DEFAULT_BOX2D: Self = Self(10.0);
+    pub const DEFAULT_BOX2D: Self = Self(Float32(10.0));
 
     /// Draw order used for 2D lines if no draw order was specified.
-    pub const DEFAULT_LINES2D: Self = Self(20.0);
+    pub const DEFAULT_LINES2D: Self = Self(Float32(20.0));
 
     /// Draw order used for 2D points if no draw order was specified.
-    pub const DEFAULT_POINTS2D: Self = Self(30.0);
+    pub const DEFAULT_POINTS2D: Self = Self(Float32(30.0));
 }
 
 impl std::cmp::PartialEq for DrawOrder {
@@ -55,6 +57,6 @@ impl Default for DrawOrder {
     #[inline]
     fn default() -> Self {
         // Pick zero as default which happens to be neither at the bottom nor the top.
-        Self(0.0)
+        Self(Float32(0.0))
     }
 }
